loader/requirements.txt
psycopg2-binary==2.9.9
hdfs==2.7.0
